chore: remove dhis-api saveEventDataValuesAndSaveEvent
