+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/program/DefaultEventService.java
@@ -27,144 +27,12 @@
  */
 package org.hisp.dhis.program;
 
-import static org.hisp.dhis.external.conf.ConfigurationKey.CHANGELOG_TRACKER;
-
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
 import lombok.RequiredArgsConstructor;
-import org.apache.commons.lang3.StringUtils;
-import org.hisp.dhis.category.CategoryOptionCombo;
-import org.hisp.dhis.category.CategoryService;
-import org.hisp.dhis.changelog.ChangeLogType;
-import org.hisp.dhis.common.IdentifiableObjectManager;
-import org.hisp.dhis.common.IllegalQueryException;
-import org.hisp.dhis.dataelement.DataElement;
-import org.hisp.dhis.eventdatavalue.EventDataValue;
-import org.hisp.dhis.external.conf.DhisConfigurationProvider;
-import org.hisp.dhis.fileresource.FileResource;
-import org.hisp.dhis.fileresource.FileResourceService;
-import org.hisp.dhis.system.util.ValidationUtils;
-import org.hisp.dhis.trackedentitydatavalue.TrackedEntityDataValueChangeLog;
-import org.hisp.dhis.trackedentitydatavalue.TrackedEntityDataValueChangeLogService;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 /**
  * @author Abyot Asalefew
  */
 @RequiredArgsConstructor
 @Service("org.hisp.dhis.program.EventService")
-public class DefaultEventService implements EventService {
-  private final IdentifiableObjectManager manager;
-
-  private final CategoryService categoryService;
-
-  private final TrackedEntityDataValueChangeLogService dataValueAuditService;
-
-  private final FileResourceService fileResourceService;
-
-  private final DhisConfigurationProvider config;
-
-  @Override
-  @Transactional
-  public void saveEventDataValuesAndSaveEvent(
-      Event event, Map<DataElement, EventDataValue> dataElementEventDataValueMap) {
-    validateEventDataValues(dataElementEventDataValueMap);
-    Set<EventDataValue> eventDataValues = new HashSet<>(dataElementEventDataValueMap.values());
-    event.setEventDataValues(eventDataValues);
-
-    event.setAutoFields();
-    if (!event.hasAttributeOptionCombo()) {
-      CategoryOptionCombo aoc = categoryService.getDefaultCategoryOptionCombo();
-      event.setAttributeOptionCombo(aoc);
-    }
-    manager.save(event);
-
-    for (Map.Entry<DataElement, EventDataValue> entry : dataElementEventDataValueMap.entrySet()) {
-      entry.getValue().setAutoFields();
-      createAndAddAudit(entry.getValue(), entry.getKey(), event);
-      handleFileDataValueSave(entry.getValue(), entry.getKey());
-    }
-  }
-
-  private String validateEventDataValue(DataElement dataElement, EventDataValue eventDataValue) {
-
-    if (StringUtils.isEmpty(eventDataValue.getStoredBy())) {
-      return "Stored by is null or empty";
-    }
-
-    if (StringUtils.isEmpty(eventDataValue.getDataElement())) {
-      return "Data element is null or empty";
-    }
-
-    if (!dataElement.getUid().equals(eventDataValue.getDataElement())) {
-      throw new IllegalQueryException(
-          "DataElement "
-              + dataElement.getUid()
-              + " assigned to EventDataValues does not match with one EventDataValue: "
-              + eventDataValue.getDataElement());
-    }
-
-    String result =
-        ValidationUtils.valueIsValid(eventDataValue.getValue(), dataElement.getValueType());
-
-    return result == null ? null : "Value is not valid:  " + result;
-  }
-
-  private void validateEventDataValues(
-      Map<DataElement, EventDataValue> dataElementEventDataValueMap) {
-    String result;
-    for (Map.Entry<DataElement, EventDataValue> entry : dataElementEventDataValueMap.entrySet()) {
-      result = validateEventDataValue(entry.getKey(), entry.getValue());
-      if (result != null) {
-        throw new IllegalQueryException(result);
-      }
-    }
-  }
-
-  private void createAndAddAudit(EventDataValue dataValue, DataElement dataElement, Event event) {
-    if (!config.isEnabled(CHANGELOG_TRACKER) || dataElement == null) {
-      return;
-    }
-
-    TrackedEntityDataValueChangeLog dataValueAudit =
-        new TrackedEntityDataValueChangeLog(
-            dataElement,
-            event,
-            dataValue.getValue(),
-            dataValue.getStoredBy(),
-            dataValue.getProvidedElsewhere(),
-            ChangeLogType.CREATE);
-
-    dataValueAuditService.addTrackedEntityDataValueChangeLog(dataValueAudit);
-  }
-
-  /** Update FileResource with 'assigned' status. */
-  private void handleFileDataValueSave(EventDataValue dataValue, DataElement dataElement) {
-    if (dataElement == null) {
-      return;
-    }
-
-    FileResource fileResource = fetchFileResource(dataValue, dataElement);
-
-    if (fileResource == null) {
-      return;
-    }
-
-    setAssigned(fileResource);
-  }
-
-  private FileResource fetchFileResource(EventDataValue dataValue, DataElement dataElement) {
-    if (!dataElement.isFileType()) {
-      return null;
-    }
-
-    return fileResourceService.getFileResource(dataValue.getValue());
-  }
-
-  private void setAssigned(FileResource fileResource) {
-    fileResource.setAssigned(true);
-    fileResourceService.updateFileResource(fileResource);
-  }
-}
+public class DefaultEventService implements EventService {}
