+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/imports/sms/EnrollmentSMSListener.java
@@ -37,6 +37,8 @@ import org.hisp.dhis.category.CategoryOptionCombo;
 import org.hisp.dhis.category.CategoryService;
 import org.hisp.dhis.common.IdentifiableObjectManager;
 import org.hisp.dhis.dataelement.DataElementService;
+import org.hisp.dhis.external.conf.DhisConfigurationProvider;
+import org.hisp.dhis.fileresource.FileResourceService;
 import org.hisp.dhis.message.MessageSender;
 import org.hisp.dhis.organisationunit.OrganisationUnit;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
@@ -64,6 +66,7 @@ import org.hisp.dhis.trackedentity.TrackedEntityType;
 import org.hisp.dhis.trackedentity.TrackedEntityTypeService;
 import org.hisp.dhis.trackedentityattributevalue.TrackedEntityAttributeValue;
 import org.hisp.dhis.trackedentityattributevalue.TrackedEntityAttributeValueService;
+import org.hisp.dhis.trackedentitydatavalue.TrackedEntityDataValueChangeLogService;
 import org.hisp.dhis.tracker.export.event.EventService;
 import org.hisp.dhis.user.User;
 import org.hisp.dhis.user.UserService;
@@ -94,8 +97,10 @@ public class EnrollmentSMSListener extends EventSavingSMSListener {
       CategoryService categoryService,
       DataElementService dataElementService,
       ProgramStageService programStageService,
-      org.hisp.dhis.program.EventService apiEventService,
       EventService eventService,
+      TrackedEntityDataValueChangeLogService dataValueAuditService,
+      FileResourceService fileResourceService,
+      DhisConfigurationProvider config,
       TrackedEntityAttributeValueService attributeValueService,
       TrackedEntityService teService,
       EnrollmentService enrollmentService,
@@ -111,8 +116,10 @@ public class EnrollmentSMSListener extends EventSavingSMSListener {
         categoryService,
         dataElementService,
         identifiableObjectManager,
-        apiEventService,
-        eventService);
+        eventService,
+        dataValueAuditService,
+        fileResourceService,
+        config);
     this.teService = teService;
     this.programStageService = programStageService;
     this.enrollmentService = enrollmentService;
