+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/imports/sms/RegisterSMSListener.java
@@ -27,23 +27,32 @@
  */
 package org.hisp.dhis.tracker.imports.sms;
 
+import static org.hisp.dhis.external.conf.ConfigurationKey.CHANGELOG_TRACKER;
+
 import java.util.Date;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
+import org.hisp.dhis.category.CategoryOptionCombo;
 import org.hisp.dhis.category.CategoryService;
+import org.hisp.dhis.changelog.ChangeLogType;
+import org.hisp.dhis.common.IdentifiableObjectManager;
+import org.hisp.dhis.common.IllegalQueryException;
 import org.hisp.dhis.dataelement.DataElement;
 import org.hisp.dhis.eventdatavalue.EventDataValue;
+import org.hisp.dhis.external.conf.DhisConfigurationProvider;
+import org.hisp.dhis.fileresource.FileResource;
+import org.hisp.dhis.fileresource.FileResourceService;
 import org.hisp.dhis.message.MessageSender;
 import org.hisp.dhis.organisationunit.OrganisationUnit;
 import org.hisp.dhis.program.Enrollment;
 import org.hisp.dhis.program.EnrollmentService;
 import org.hisp.dhis.program.EnrollmentStatus;
 import org.hisp.dhis.program.Event;
-import org.hisp.dhis.program.EventService;
 import org.hisp.dhis.program.UserInfoSnapshot;
 import org.hisp.dhis.sms.command.SMSCommand;
 import org.hisp.dhis.sms.command.code.SMSCode;
@@ -51,26 +60,41 @@ import org.hisp.dhis.sms.incoming.IncomingSms;
 import org.hisp.dhis.sms.incoming.IncomingSmsService;
 import org.hisp.dhis.sms.incoming.SmsMessageStatus;
 import org.hisp.dhis.sms.listener.CommandSMSListener;
+import org.hisp.dhis.system.util.ValidationUtils;
+import org.hisp.dhis.trackedentitydatavalue.TrackedEntityDataValueChangeLog;
+import org.hisp.dhis.trackedentitydatavalue.TrackedEntityDataValueChangeLogService;
 import org.hisp.dhis.user.CurrentUserUtil;
 import org.hisp.dhis.user.UserService;
 
 @Slf4j
 public abstract class RegisterSMSListener extends CommandSMSListener {
 
-  protected final EventService eventService;
-
   protected final EnrollmentService enrollmentService;
+
+  protected final TrackedEntityDataValueChangeLogService dataValueAuditService;
+
+  protected final FileResourceService fileResourceService;
+
+  protected final DhisConfigurationProvider config;
+
+  protected final IdentifiableObjectManager identifiableObjectManager;
 
   protected RegisterSMSListener(
       CategoryService dataElementCategoryService,
       UserService userService,
       IncomingSmsService incomingSmsService,
       MessageSender smsSender,
-      EventService eventService,
-      EnrollmentService enrollmentService) {
+      EnrollmentService enrollmentService,
+      TrackedEntityDataValueChangeLogService dataValueAuditService,
+      FileResourceService fileResourceService,
+      DhisConfigurationProvider config,
+      IdentifiableObjectManager identifiableObjectManager) {
     super(dataElementCategoryService, userService, incomingSmsService, smsSender);
-    this.eventService = eventService;
     this.enrollmentService = enrollmentService;
+    this.dataValueAuditService = dataValueAuditService;
+    this.fileResourceService = fileResourceService;
+    this.config = config;
+    this.identifiableObjectManager = identifiableObjectManager;
   }
 
   protected void register(
@@ -133,7 +157,7 @@ public abstract class RegisterSMSListener extends CommandSMSListener {
       }
     }
 
-    eventService.saveEventDataValuesAndSaveEvent(event, dataElementsAndEventDataValues);
+    saveEventDataValuesAndSaveEvent(event, dataElementsAndEventDataValues);
 
     update(sms, SmsMessageStatus.PROCESSED, true);
 
@@ -141,5 +165,105 @@ public abstract class RegisterSMSListener extends CommandSMSListener {
         StringUtils.defaultIfEmpty(smsCommand.getSuccessMessage(), SMSCommand.SUCCESS_MESSAGE),
         sms.getOriginator(),
         INFO);
+  }
+
+  public void saveEventDataValuesAndSaveEvent(
+      Event event, Map<DataElement, EventDataValue> dataElementEventDataValueMap) {
+    validateEventDataValues(dataElementEventDataValueMap);
+    Set<EventDataValue> eventDataValues = new HashSet<>(dataElementEventDataValueMap.values());
+    event.setEventDataValues(eventDataValues);
+
+    event.setAutoFields();
+    if (!event.hasAttributeOptionCombo()) {
+      CategoryOptionCombo aoc = dataElementCategoryService.getDefaultCategoryOptionCombo();
+      event.setAttributeOptionCombo(aoc);
+    }
+    identifiableObjectManager.save(event);
+
+    for (Map.Entry<DataElement, EventDataValue> entry : dataElementEventDataValueMap.entrySet()) {
+      entry.getValue().setAutoFields();
+      createAndAddAudit(entry.getValue(), entry.getKey(), event);
+      handleFileDataValueSave(entry.getValue(), entry.getKey());
+    }
+  }
+
+  private String validateEventDataValue(DataElement dataElement, EventDataValue eventDataValue) {
+
+    if (StringUtils.isEmpty(eventDataValue.getStoredBy())) {
+      return "Stored by is null or empty";
+    }
+
+    if (StringUtils.isEmpty(eventDataValue.getDataElement())) {
+      return "Data element is null or empty";
+    }
+
+    if (!dataElement.getUid().equals(eventDataValue.getDataElement())) {
+      throw new IllegalQueryException(
+          "DataElement "
+              + dataElement.getUid()
+              + " assigned to EventDataValues does not match with one EventDataValue: "
+              + eventDataValue.getDataElement());
+    }
+
+    String result =
+        ValidationUtils.valueIsValid(eventDataValue.getValue(), dataElement.getValueType());
+
+    return result == null ? null : "Value is not valid:  " + result;
+  }
+
+  private void validateEventDataValues(
+      Map<DataElement, EventDataValue> dataElementEventDataValueMap) {
+    String result;
+    for (Map.Entry<DataElement, EventDataValue> entry : dataElementEventDataValueMap.entrySet()) {
+      result = validateEventDataValue(entry.getKey(), entry.getValue());
+      if (result != null) {
+        throw new IllegalQueryException(result);
+      }
+    }
+  }
+
+  private void createAndAddAudit(EventDataValue dataValue, DataElement dataElement, Event event) {
+    if (!config.isEnabled(CHANGELOG_TRACKER) || dataElement == null) {
+      return;
+    }
+
+    TrackedEntityDataValueChangeLog dataValueAudit =
+        new TrackedEntityDataValueChangeLog(
+            dataElement,
+            event,
+            dataValue.getValue(),
+            dataValue.getStoredBy(),
+            dataValue.getProvidedElsewhere(),
+            ChangeLogType.CREATE);
+
+    dataValueAuditService.addTrackedEntityDataValueChangeLog(dataValueAudit);
+  }
+
+  /** Update FileResource with 'assigned' status. */
+  private void handleFileDataValueSave(EventDataValue dataValue, DataElement dataElement) {
+    if (dataElement == null) {
+      return;
+    }
+
+    FileResource fileResource = fetchFileResource(dataValue, dataElement);
+
+    if (fileResource == null) {
+      return;
+    }
+
+    setAssigned(fileResource);
+  }
+
+  private FileResource fetchFileResource(EventDataValue dataValue, DataElement dataElement) {
+    if (!dataElement.isFileType()) {
+      return null;
+    }
+
+    return fileResourceService.getFileResource(dataValue.getValue());
+  }
+
+  private void setAssigned(FileResource fileResource) {
+    fileResource.setAssigned(true);
+    fileResourceService.updateFileResource(fileResource);
   }
 }
