+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/imports/sms/SimpleEventSMSListener.java
@@ -35,6 +35,8 @@ import org.hisp.dhis.category.CategoryOptionCombo;
 import org.hisp.dhis.category.CategoryService;
 import org.hisp.dhis.common.IdentifiableObjectManager;
 import org.hisp.dhis.dataelement.DataElementService;
+import org.hisp.dhis.external.conf.DhisConfigurationProvider;
+import org.hisp.dhis.fileresource.FileResourceService;
 import org.hisp.dhis.message.MessageSender;
 import org.hisp.dhis.organisationunit.OrganisationUnit;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
@@ -54,6 +56,7 @@ import org.hisp.dhis.smscompression.models.SmsSubmission;
 import org.hisp.dhis.smscompression.models.Uid;
 import org.hisp.dhis.trackedentity.TrackedEntityAttributeService;
 import org.hisp.dhis.trackedentity.TrackedEntityTypeService;
+import org.hisp.dhis.trackedentitydatavalue.TrackedEntityDataValueChangeLogService;
 import org.hisp.dhis.tracker.export.event.EventService;
 import org.hisp.dhis.user.User;
 import org.hisp.dhis.user.UserService;
@@ -76,8 +79,10 @@ public class SimpleEventSMSListener extends EventSavingSMSListener {
       OrganisationUnitService organisationUnitService,
       CategoryService categoryService,
       DataElementService dataElementService,
-      org.hisp.dhis.program.EventService apiEventService,
       EventService eventService,
+      TrackedEntityDataValueChangeLogService dataValueAuditService,
+      FileResourceService fileResourceService,
+      DhisConfigurationProvider config,
       EnrollmentService enrollmentService,
       IdentifiableObjectManager identifiableObjectManager) {
     super(
@@ -91,8 +96,10 @@ public class SimpleEventSMSListener extends EventSavingSMSListener {
         categoryService,
         dataElementService,
         identifiableObjectManager,
-        apiEventService,
-        eventService);
+        eventService,
+        dataValueAuditService,
+        fileResourceService,
+        config);
     this.enrollmentService = enrollmentService;
   }
